moved things from Home to AnnualReport + refactored Annual Report to look like prototype (WIP)
