+++ coops-ui/src/components/AnnualReport.vue
@@ -5,30 +5,46 @@
     <div class="loading-container fade-out">
       <div class="loading__content">
         <v-progress-circular color="primary" :size="50" indeterminate></v-progress-circular>
-        <div class="loading-msg">Preparing Your Annual Report</div>
+        <div class="loading-msg">Preparing Your {{ year }} Annual Report</div>
       </div>
     </div>
 
-    <v-container class="view-container">
-      <article id="annual-report-article">
-        <h1 id="AR-header">{{year}} Annual Report - <span style="font-style: italic">{{ reportState }}</span></h1>
-        <p>Select your Annual General Meeting (AGM) date, and verify or change your Registered office address and List
-          of Directors as of your AGM.</p>
+    <v-container id="annual-report-container" class="view-container">
+      <article id="annual-report-article" :class="this.agmDate ? 'agm-date-selected':'no-agm-date-selected'">
+        <header>
+          <h1 id="AR-header">File {{ year }} Annual Report -
+            <span style="font-style: italic">{{ reportState }}</span>
+          </h1>
+          <p>Select your Annual General Meeting (AGM) date, and verify or change your Registered office address and List
+            of Directors as of your AGM.</p>
+        </header>
+
         <div v-if="filedDate == null">
+          <!-- Annual General Meeting Date -->
           <section>
-            <h3 id="AR-step-1-header">1. Enter your Annual General Meeting Date</h3>
-            <div id="AR-step-1-container">
-              <AGMDate/>
-            </div>
+            <header>
+              <h2 id="AR-step-1-header">1. Annual General Meeting Date</h2>
+            </header>
+            <v-card flat id="AR-step-1-container">
+              <AGMDate ref="ARFilingDate" v-on:childToParent="onChildClick"/>
+            </v-card>
           </section>
+
+          <!-- Addresses -->
           <section>
-            <h3 id="AR-step-2-header">2. Registered Office Addresses</h3>
-            <div id="AR-step-2-container">
+            <header>
+              <h2 id="AR-step-2-header">2. Registered Office Addresses
+                <span class="agm-date">(as of {{ year }} Annual General Meeting)</span>
+              </h2>
+              <p>Verify or change your Registered Office Addresses.</p>
+            </header>
+            <v-card flat id="AR-step-2-container">
               <RegisteredOfficeAddress/>
-            </div>
+            </v-card>
           </section>
 
           <!-- Director Information -->
+          <!-- TODO - should be in section as above -->
           <Directors ref="directors" @directorsChange="directorsChangeEventHandler" />
 
         </div>
@@ -36,16 +52,29 @@
           <ARComplete/>
         </div>
       </article>
+
       <aside>
         <affix relative-element-selector="#annual-report-article" :offset="{ top: 120, bottom: 40 }">
           <sbc-fee-summary v-bind:filingData="[...filingData]" />
         </affix>
       </aside>
     </v-container>
+
+     <v-container id="submit-container" class="pt-0">
+      <div class="ar-filing-buttons">
+        <v-btn v-if="filedDate == null" id="ar-pay-btn" color="primary" large :disabled="!validated"
+          @click="submit">File & Pay</v-btn>
+        <v-btn v-else id="ar-next-btn" color="primary" large :disabled="currentYear == ARFilingYear"
+          @click="nextAR">Next</v-btn>
+        <v-btn large to="/">Cancel</v-btn>
+      </div>
+    </v-container>
+
   </div>
 </template>
 
 <script>
+import axios from '../axios-auth'
 import AGMDate from '@/components/ARSteps/AGMDate.vue'
 import RegisteredOfficeAddress from '@/components/ARSteps/RegisteredOfficeAddress.vue'
 import Directors from '@/components/ARSteps/Directors.vue'
@@ -65,6 +94,7 @@ export default {
   },
   data () {
     return {
+      agmDate: '',
       directorsChange: false,
       filingData: []
     }
@@ -82,6 +112,15 @@ export default {
     filedDate () {
       return this.$store.state.filedDate
     },
+    validated () {
+      return this.$store.state.validated
+    },
+    currentYear () {
+      return this.$store.state.currentDate.substring(0, 4)
+    },
+    ARFilingYear () {
+      return this.$store.state.ARFilingYear
+    },
     year () {
       return this.$store.state.ARFilingYear
     },
@@ -93,11 +132,58 @@ export default {
   mounted () {
   },
   methods: {
+    onChildClick (value) {
+      this.agmDate = value
+    },
     directorsChangeEventHandler (val) {
       this.directorsChange = val
     },
     getDirectors () {
       this.$refs.directors.getDirectors()
+    },
+    submit () {
+      // todo: redirect to payment - will need to save state of page
+      var token = sessionStorage.getItem('KEYCLOAK_TOKEN')
+      // probably need to parametrize date=this.$store.state.currentDate + add token in header for api
+      var url = this.payURL
+      var paymentJson
+      // other team doing credit card entering/payment confirmation? - don't know what to check for in result
+      axios.get(url).then(response => {
+        paymentJson = response.data
+        console.log('payment response: ', paymentJson)
+        if (paymentJson) this.$store.state.filedDate = this.$store.state.currentDate
+      }).catch(error => console.log('payment ERROR: ' + error))
+      this.$store.state.filedDate = this.$store.state.currentDate
+      console.log('submit: filedDate =', this.$store.state.filedDate)
+    },
+    resetARInfo () {
+      this.$store.state.agmDate = null
+      this.$store.state.filedDate = null
+      this.$store.state.validated = false
+      this.$store.state.noAGM = false
+      this.$store.state.regOffAddrChange = false
+      this.setRegOffAddrNull()
+      this.$router.go()
+    },
+    nextAR () {
+      this.resetARInfo()
+      this.getARInfo(this.corpNum)
+    },
+    setRegOffAddrNull () {
+      this.$store.state.DeliveryAddressStreet = null
+      this.$store.state.DeliveryAddressStreetAdditional = null
+      this.$store.state.DeliveryAddressCity = null
+      this.$store.state.DeliveryAddressRegion = null
+      this.$store.state.DeliveryAddressPostalCode = null
+      this.$store.state.DeliveryAddressCountry = null
+      this.$store.state.DeliveryAddressInstructions = null
+      this.$store.state.MailingAddressStreet = null
+      this.$store.state.MailingAddressStreetAdditional = null
+      this.$store.state.MailingAddressCity = null
+      this.$store.state.MailingAddressRegion = null
+      this.$store.state.MailingAddressPostalCode = null
+      this.$store.state.MailingAddressCountry = null
+      this.$store.state.MailingAddressInstructions = null
     },
     toggleFiling (setting, filing) {
       var added = false
@@ -147,55 +233,23 @@ export default {
 }
 </script>
 
-<style lang="stylus">
+<style lang="stylus" scoped>
   @import "../assets/styles/theme.styl"
 
-  section
-    header p
-      color $gray6
-      font-size 1rem
-
-  .view-container
-    display flex
-    flex-flow column nowrap
-    padding-top 3rem
-    padding-bottom 3rem
-
   article
-    flex 1 1 auto
+    .v-card
+      line-height 1.2rem
+      font-size 0.875rem
 
-  aside
-    flex 0 0 auto
-    width 20rem
+  section p
+    //font-size 0.875rem
+    color $gray6
+
+  section + section
     margin-top 3rem
-    margin-right -8rem
 
-    .affix
-      width 20rem
-
-  @media (min-width 960px)
-    .view-container
-      flex-flow row nowrap
-
-      article
-        margin-right 2rem
-
-      aside
-        margin-top 0
-        width 20rem
-
-  @media (max-width 768px)
-    .view-container
-      aside
-        width 100%
-
-        .affix
-          position relative
-          top 0 !important
-          width 100%
-
-  #annual-report
-    width 100%
+  h2
+    margin-bottom 0.25rem
 
   #AR-header
     margin-bottom 1.25rem
@@ -217,7 +271,15 @@ export default {
   .title-container
     margin-bottom 0.5rem
 
-  p
-    margin-top 1rem
+  .agm-date
+    margin-left 0.25rem
+    font-weight 300
 
+  // Filing Buttons
+  .ar-filing-buttons
+    padding-top 2rem
+    border-top: 1px solid $gray5
+    text-align right
+    .v-btn + .v-btn
+      margin-left 0.5rem
 </style>
